pyproject.toml
[tool.poetry]
name = "pypyrus"
version = "0.1.0"
description = ""
authors = ["Ikarys <bikarys@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
pydantic = "^2.6.3"
icecream = "^2.1.3"
questionary = "^2.0.1"
jinja2 = "^3.1.3"
loguru = "^0.7.2"
types-pyyaml = "^6.0.12.12"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest = "^8.0.2"
vulture = "^2.11"
bandit = "^1.7.8"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[mypy]
python_version = 3.11
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = ["--cov=src", "--import-mode=importlib"]
testpaths = ["tests/"]
pythonpath = "src"